Allow for multiple Tenant ID's to be included
In some cases, we've seen the requirement for allowing multiple tenants to be included on the same platform.
Since it's already allowed to use addTenants with an array, it makes sense, that it should also scope by multiple ID's. 

Could be extended to allow for a collection object, instead of only arrays if it has potential.